requirements.txt
psycopg2-binary==2.9.9
fastapi==0.104.0
pydantic==2.4.2
pytest==7.4.0
uvicorn==0.23.2
httpx==0.27.2